Move some expression data type propagation calculations to a separate pkg
